json scheme change

+++ app/src/main/java/cz/matej/app/strvacademyweather/entity/openweather/Wind.java
@@ -12,7 +12,7 @@ public class Wind
 	private Double speed;
 	@SerializedName("deg")
 	@Expose
-	private Integer deg;
+	private Double deg;
 
 
 	/**
@@ -27,7 +27,7 @@ public class Wind
 	 * @param speed
 	 * @param deg
 	 */
-	public Wind(Double speed, Integer deg)
+	public Wind(Double speed, Double deg)
 	{
 		this.speed = speed;
 		this.deg = deg;
@@ -55,7 +55,7 @@ public class Wind
 	/**
 	 * @return The deg
 	 */
-	public Integer getDeg()
+	public Double getDeg()
 	{
 		return deg;
 	}
@@ -64,7 +64,7 @@ public class Wind
 	/**
 	 * @param deg The deg
 	 */
-	public void setDeg(Integer deg)
+	public void setDeg(Double deg)
 	{
 		this.deg = deg;
 	}
